# This is a combination of 2 commits. # This is the 1st commit message:
Workaround compiler limitation

The compiler is unable to reason that the immutable borrow does not need
to live inside the `None` branch, and thinks that the mutable borrow
will conflict with it.

The workaround implemented here was to load the value twice, so that the
first one has a shorter lifetime.

# The commit message #2 will be skipped:

# RocksDB multicolumn support