Refactor `send_recv_info_bytes` to not use bytes
Receive a `SerializedMessage` as a parameter and serialize it before
forwarding it to `send_recv_bytes_internal`. The method was renamed to
remove the `_bytes` suffix and usages were simplified because
serialization was no longer necessary to be done by the caller.

PR #42: Implement a `Codec` for the network protocol