Fix checking if bucket was newly created
Refactor the `try_create_bucket` method into `create_bucket_if_needed`
instead, and make it only try to create the bucket after performing a
check to see if the bucket already exists.

This fixes an issue where the bucket already existed, but the call to
create the bucket again did not do anything nor return an error.