Fix incorrect loading of certificate keys
A call to the certificate `MapView`'s `indices` method lead to an error
when it incorrectly thought that chain state sub-views were entries in
the map.

This happened because although the chain states and the certificates
were kept separate by the `BaseKey` type, the `MapView` still used the
same context as a hypothetical `Collection` of `ChainStateView`s. That
meant that the `MapView` thought that `BaseKey::ChainState` keys were
also owned by the map, even though all certificates reads and writes
used the `BaseKey::Certificate` variant.

The solution was to change the `MapView` to use a separate context with
a separate key prefix. The `BaseKey::Certificate` variant was changed to
have the certificate hash dropped and instead is used for the whole
`MapView`. That allowed changing the `MapView` to not need to use
`BaseKey` as its key type, using `HashValue` directly instead.