Second restructuration of the view code (#189)
* First step of cleanup.

* Removal the Batch from the common code.

* Partial cleanup of the derive_key exception handling.

* Move the type code in another place.

* Eliminate the create_batch that is not needed right now.

* Moving away the put_item_batch / remove_item_batch.

* Removal of the run_with_batch code.

* Move the write_batch as part of the KeyValueOperations.

* Renaming of the db variable.

* Put the dynamo_db as KeyValueOperations.

* Restructuration of the Memory code for KeyValueOperations.

* Put the dependent type of error.

* Remove the write_key that had a single usage.

* Renaming of the main containing type with postfix Container.

* Some attempt at implementing the Context with macros.

* Removal of unused code and debugging of the macro.

* reformatting.

* Implementation using the type alias and derive_bounded.

* Restructuration of the batch stuff.

* Some renaming.

* Simplification of the template structure.

* Remove unused code.

* Apply reformatting.

* Update linera-views/src/macros.rs

Co-authored-by: Janito Vaqueiro Ferreira Filho <janito.vff@gmail.com>

* Update linera-views/src/macros.rs

Co-authored-by: Janito Vaqueiro Ferreira Filho <janito.vff@gmail.com>

* Correction for compilation.

* Create initial `linera-sdk` crate

Create the empty crate.

* Create a `HashValue` type

Equivalent to `linera_base::crypto::HashValue`, but without the need for
the `generic_array` dependency.

* Copy some types from `linera_base::messages`

The types that are required inside a contract.

* Document `ApplicationId` type

Add a short summary to describe the type.

* Copy some types from `linera-execution`

Copy the types necessary inside an application's implementation.

Derive `Default` for `RawExecutionResult` instead of copying the custom
`Default` implementation because the plan is to remove the generic type
parameter (and therefore the inferred generic constraint) later.

* Rename `NewSession` to `Session`

Allow the type to be reused in different scenarios that isn't just
session creation.

* Rename `RawExecutionResult` into `ExecutionResult`

Remove the generic type parameter and rename the type because there's no
need to differentiate from a system execution result.

* Create an `Application` trait

Equivalent to `linera_execution::UserApplication` but without the
storage context parameters. That's because the assumption is that `self`
will contain the application's state loaded from storage.

* Create an `ExportedFuture` helper type

Implement a way to make host futures callable from guest WASM modules.

* Define the application interface in a WIT file

Make it possible to generate the WASM application interface more easily.

* Define the interface exported to an application

Create a WIT file with the interface that an application can use.

* Implement conversions for custom `HashValue` type

Make it easier to convert to and from the representation used in the
WASM interface.

* Replace `usize` with `u64`

Ensure the value is the same size on all platforms.

* Create a simple counter application example

A trivially simple application implementation. Supports operations and
cross-application calls that increment the counter by a value, and
supports queries to get the counter value.

* Create a `boilerplate` module

The goal is to allow this module to be shared between all application
examples, at least until the SDK is able to generate the boilerplate
code.

* Import system interface

Will be used to handle the application state.

* Handle application state management

Write code to make it simpler to access the application state. To make
the code generic over different application examples, an
`ApplicationState` type alias is assumed to exist in the example.

* Export stubbed application endpoints

Create stubbed versions of the types required by the code generated from
`wit-bindgen`.

* Add more conversions from WIT types

Implement conversions for the types used as parameters in an
application's public endpoints.

* Implement conversions to WIT types

Implement conversions for types the application endpoints are expected
to return.

* Finish implementing exported futures

Load the application state (locking it for writes if necessary) and call
the respective method in the application type that implements
`linera_sdk::Application`.

* Add CI job to build the WASM application examples

Ensure `linera-sdk` builds correctly for the WASM target.

* refactor: chain errors + worker errors (#173)

Co-authored-by: Janito Vaqueiro Ferreira Filho <janito.vff@gmail.com>
Co-authored-by: Mathieu Baudet <1105398+ma2bd@users.noreply.github.com>

* Update minimum supported Rust version to 1.65.0

Enable support for `Backtrace` and generic associated types.

* Update CI workflows to use new Rust version

Ensure CI uses the minimum supported Rust version.

* Remove unused borrows

Suggested by a Clippy lint.

* Use renamed `clippy::disallowed_names` attribute

Update the attribute to use the new Clippy lint name.

* Ignore large error type Clippy lints for now

The error types are currently being refactored, so this new Clippy lint
will be applied later.

* fix old renaming from Zef to Linera

* chore: pre-push git hook installation script (#187)

* some suggested fixes

* Code restructuration.

Co-authored-by: Mathieu Dutour Sikiric <mathieu.dutour@linera.io>
Co-authored-by: Janito Vaqueiro Ferreira Filho <janito.vff@gmail.com>
Co-authored-by: Janito Vaqueiro Ferreira Filho <janito.vff@linera.io>
Co-authored-by: Christos Hadjiaslanis <christos.hadjiaslanis@linera.io>
Co-authored-by: Mathieu Baudet <1105398+ma2bd@users.noreply.github.com>