Store published bytecode IDs in the chain state
Map published bytecode IDs to their location. This effect has to be
applied immediately when it's received, so that the bytecode's
certificate is known.

It can't be part of the execution state because bytecodes could be
received at different heights on different validators, which would break
consensus on the execution state hash.