Start implementing transfers initiated by users
Handle application operations as signed transfers that may credit an
account on the current micro-chain or create an effect to credit an
account on a different micro-chain.

The transfer is only executed if the payload is correctly signed. This
is still not enough for security, because there is no attempt to prevent
replay attacks. The next commits will add more checks to enforce
security.