Expose ExecuteOperation API through GraphQL mutations
We have a core system API which is expressed via the `SystemOperation` enum.
This system API needs to exposed to the outside world so that it can be
easily consumed programatically.

A set of GraphQL mutations have been created which correspond directly to each
operation, or variant of the SystemOperation enum, and the requisite types have
been modified to behave as GraphQL primitives.