Allow adding messages to blocks
Add a `BlockBuilder::with_message` that allows adding a message by
referencing it using the `EffectId` that sent the message. For this to
work, the message has to be obtained from the validator's worker state,
and that requires asynchronous code. Therefore, the new method simply
queues the method to be fetched later when the block is being signed.