Fix some commands so they work with non-owned chains.
Some commands panicked if there were any chains that we don't have the
private key to.

This also updates the reconfiguration test so that the node service has
the key and can receive new committees. That means the other client does
not have the key to that chain when it uses those commands.