Move round number to Certificate.
It's possible that two `ConfirmedBlock` certificates for the same block
with different round numbers get a quorum of signatures. To make sure
validators agree on the correct `previous_block_hash` for the next
block, the hash must not depend on the round number.

This commit moves it from `CertificateValue` to `Certificate` (and
`Vote` etc.). The signatures still apply to the value _and_ the round
number.