Fix liveness issue for multi owner chain manager.
Assume there are four equally-weighted validators, A, B, C, and D.
Then for safety purposes, one faulty validator is tolerated.

This change addresses the following liveness issue:

In round 0, block X is proposed, and all validators vote to validate it,
so there is a `ValidatedBlock` certificate. However, it only gets sent
to A at first, who votes to confirm it, so A has locked block X.

The other validators first see block Y proposed in round 1. Since they
haven't locked any block yet, they vote to validate it. Even if they
later receive the `ValidatedBlock` certificate from round 0, they now
won't vote to confirm it anymore. The three of them create a certificate
for a `ValidatedBlock`.

Now faulty validator D goes offline, and block X is proposed in round 2.
Validator A votes to validated X in round 2, and therefore can't vote
to confirm Y in 1 anymore.

That effectively blocks the chain, since A cannot vote for any proposal
different from X anymore, B and C can only vote for Y, and D is gone.

The fix is to allow A to vote for Y.

In general, if there is a `ValidatedBlock` certificate from a round
higher than the locked block's, voting for it is allowed.