Fix internal alignment in `frunk` heterogeneous lists (#929)
* Make some `GuestPointer` methods `const`

Allow them to be used in the `const` context that calculates the size of
a heterogeneous list.

* Add comment explaining how padding is calculated

Link to the Wikipedia source.

* Test `GuestPointer::align_at`

Ensure that the non-trivial padding calculation is correct.

* Fix alignment between heterogeneous list elements

Because heterogeneous lists are recursive types, some special handling
is needed. Alignment between internal elements must be calculated
separately, because otherwise the alignment for the whole list (the
tail) is used, which is not the same as the alignment as the next
element.

The same is true when reading and writing to memory, because the
alignment between each element must be considered, and not the alignment
for all the remaining elements.

* Fix documentation for test type

There's only one place where padding is needed.

* Add roundtrip tests for some `hlist` types

Make sure that they can be stored in memory and read from memory, and
that they can be lowered into its flat layout and lifted back from it.
Ensure that their internal element alignment is properly respected.