Add `wit_export` attribute macro to export host functions to Wasm modules (#955)
* Export `wit_import` macro from Witty

Remove the requirement to add a direct dependency to
`linera-witty-macros`.

* Create an `ExportTo` helper trait

The trait that the `wit_export` attribute macro will implement,
registering host functions as exports to a guest Wasm instance.

* Create an `ExportFunction` trait

Allows runtimes to determine how host functions are exported to guest
Wasm instances.

* Represent different ways to return host values

Create a `ResultStorage` trait in order to represent how a host
function's output needs to be sent to the guest Wasm module.

* Implement how guest instances see host functions

Create a `GuestInterface` trait to help map how a flattened host
function's signature maps to how a guest Wasm instance interfaces with
the host function.

* Create an `ExportedFunctionInterface` helper trait

Use the `GuestInterface` and `ResultStorage` traits in a single trait
that can be used to determine the interface to an exported function from
the host to the guest Wasm instance.

* Add Wasmer support for exporting host functions

Implement `ExportFunction` for the function signatures supported by
Wasmer.

* Add Wasmtime support for exporting host functions

Implement `ExportFunction` for the function signatures supported by
Wasmtime.

* Extract information from a function to export

Collect the necessary pieces to be used in the code generation for
exported functions.

* Add a `wit_export` attribute procedural macro

The macro will implement `ExportTo` to different Wasm runtime-dependent
types, depending on which feature flags are enabled. For now no runtimes
are supported.

* Add Wasmer support for `wit_export`

Generate code to export host functions to guest Wasm instances using
Wasmer.

* Add Wasmtime support for `wit_export`

Generate code to export host functions to guest Wasm instances using
Wasmtime.

* Refactor to add module group to `TestInstance`

Prepare to have separate test modules for `wit_export` tests.

* Prepare to allow exporting host functions in tests

Allow providing an initializer function to the `load_test_module` method
that allows exporting host functions to a guest Wasm instance builder.

* Add test Wasm modules for `wit_export`

Add some test modules that import host functions.

* Test `wit_export` with Wasmer and Wasmtime

Check that the exported host functions are called correctly.

* Support exported functions in `MockInstance`

Implement `ExportFunction` for `MockInstance` in order to allow testing
exporting functions into the `MockInstance`, and add a helper method to
allow calling the exported functions.

* Mock the test modules for the `wit_export` test

Prepare to test `wit_export` using a `MockInstance`.

* Create a `MockResults` helper trait

Make it easier to write generic tests.

* Add `MockInstance` support for `wit_export` macro

Generate code that can be used in tests with a `MockInstance`.

* Test `wit_export` using a `MockInstance`

Provide some integration tests that don't require a Wasm runtime to run.