Replace `Self` with the type name
Prepare to be able to call the methods with possibly different lifetime
parameters.

This is needed because if the type has a generic type parameter for the
caller, which requires a lifetime, then using `Self` requires the
lifetime to always be the same. That's not possible because the caller
only lives inside a closure, and there's no way to link the two
lifetimes. Replacing `Self` with the type name allows the lifetime
parameter to change.