Simplify generics for exporting reentrant functions with Witty (#965)
* Require caller parameter to be a mutable reference

Better match the semantics required by the generated code.

* Don't fail `wit_export` on types with generics

Prepare to allow specifying the generic caller type parameter in the
type.

* Refactor to remove `From` implementation

Prepare to receive an extra parameter in the constructors.

* Replace `Self` with the type name

Prepare to be able to call the methods with possibly different lifetime
parameters.

This is needed because if the type has a generic type parameter for the
caller, which requires a lifetime, then using `Self` requires the
lifetime to always be the same. That's not possible because the caller
only lives inside a closure, and there's no way to link the two
lifetimes. Replacing `Self` with the type name allows the lifetime
parameter to change.

* Add caller generic type parameter to target type

Detect if the target type is using a generic type parameter for the
caller, and fill it appropriately when generating code.

* Detect reentrant functions sharing caller types

If the generic type parameter for the caller type is defined in the
`impl` block, forward it to the `FunctionInformation` so that it can
detect if it's used as the first function parameter to determine if the
function is reentrant or not.

* Add `TestInstance::Caller` associated type

Map the test instances to the caller types they use. This is used to
allow the tests to use interface types with generic caller type
parameters.

* Share the generic caller type parameter in tests

Avoid repeating the trait constraints in every reentrant function.