Updating local Kubernetes CI (#1062)
## Motivation

We have Kubernetes deployments that we're using now, so we need to have a CI to make sure we're not breaking the local run

## Proposal

Update old Kubernetes CI to do what we do in the current local deployment

## Test Plan

CI

## Release Plan

<!--
How to safely release the changes. Please create issues to track future release work.
-->
- [x] All good!