Ensure streams in `select_all` can be paused
The `SelectAll` future will drop the `stream::Next` futures, so the
streams stop being polled temporarily while `update_streams` is called.
That means that the deadlock can occur if the streams are holding a lock
that `update_streams` use.

The solution here is to ensure that the stream continues executing while
`update_streams` is running, so that any locks that are held are
eventually released.