Switch GitHub actions to actions-rust-lang/setup-toolchain
Removes explicit reference to Swatinem/rust-cache as
actions-rust-lang/setup-rust-toolchain automatically uses this action
to enable caching (unless we set `cache: false`).

Also switches our toolchain environment to be the same as the build
environment, as the `rust-toolchain.toml` file overrides it anyway.
This is undesirable: nightly rustfmt has some additional options we
would like to keep.