Copy existing binaries to Docker image (#1188)
## Motivation

Development cycles are slow when deploying a local cluster because of how slow the Docker build is, which also slows down CI.

## Proposal

Let's have an option to build locally the binaries (outside of Docker) then just copy them over to build the image. This should speed up development cycles for people using Linux. AFAIK there are no Apple M1 Silicon Docker images that we could use, so people not using Linux (me) are out of luck on this.

## Test Plan

This will be triggered in the following PR by CI, which we'll use as a test for this (I can't test it locally because I don't have a Linux machine)