Switch GitHub actions to actions-rust-lang/setup-toolchain (#1182)
* Switch GitHub actions to actions-rust-lang/setup-toolchain

Removes explicit reference to Swatinem/rust-cache as
actions-rust-lang/setup-rust-toolchain automatically uses this action
to enable caching (unless we set `cache: false`).

Also switches our toolchain environment to be the same as the build
environment, as the `rust-toolchain.toml` file overrides it anyway.
This is undesirable: nightly rustfmt has some additional options we
would like to keep.

* Use my branch to allow overriding the toolchain file

Since we use nightly Rust tooling for our linting, we want to override
our usual toolchain for the linting step.  Upstream
`setup-rust-toolchain` currently doesn't allow this.  I have submitted
a PR at
https://github.com/actions-rust-lang/setup-rust-toolchain/pull/26 to
add the functionality, but until then this commit points to a branch
on my personal repository at
https://github.com/Twey/setup-rust-toolchain/.  We should revert it
once the PR is merged.