`linera-views-derive`: derive filterable collections in `GraphQLView`
Instead of a function from key to value, represent `CollectionView`s
in GraphQL as an object with `value`, which behaves identically to the
previous function, and `values`, which takes a structured input
object as parameters and returns a list of all values matching the
parameters.

Currently, the input only supports filtering by a list of keys, but it
could be extended in the future to support more sophisticated
filtering predicates as well as pagination of results.