`linera-views-derive`: Use `insta` for snapshot tests
[`insta`](https://insta.rs/) is a tool for automatically taking
snapshots of correct output and using them as the basis for tests that
said output hasn't changed.

The snapshots themselves are stored next to the test, as `.snap` files
in a `snapshots` directory.  If a snapshot test breaks, it will create
a `.snap.new` file containing the new output; if, on inspection, the
changes to this file match your expectations, you can remove the
`.new` suffix to promote it to presumed-correct output.
Alternatively, the command-line tool `cargo-insta` provides a more
streamlined flow.