Codify linting toolchain in a toolchain file
Moves the nightly toolchain we use for linting out of a GitHub action
and into a dedicated `toolchains/lint/rust-toolchain.toml` file.

rustup doesn't currently support specifying a toolchain file, though
[it will soon](https://github.com/rust-lang/rustup/issues/2686), so we
hack around it by moving the toolchain file back and forth as part of
the GitHub action for Rust.