Filterable map views (#1237)
* `linera-views-derive`: derive filterable collections in `GraphQLView`

Instead of a function from key to value, represent `CollectionView`s
in GraphQL as an object with `value`, which behaves identically to the
previous function, and `values`, which takes a structured input
object as parameters and returns a list of all values matching the
parameters.

Currently, the input only supports filtering by a list of keys, but it
could be extended in the future to support more sophisticated
filtering predicates as well as pagination of results.

* `linera-views-derive`: Use `insta` for snapshot tests

[`insta`](https://insta.rs/) is a tool for automatically taking
snapshots of correct output and using them as the basis for tests that
said output hasn't changed.

The snapshots themselves are stored next to the test, as `.snap` files
in a `snapshots` directory.  If a snapshot test breaks, it will create
a `.snap.new` file containing the new output; if, on inspection, the
changes to this file match your expectations, you can remove the
`.new` suffix to promote it to presumed-correct output.
Alternatively, the command-line tool `cargo-insta` provides a more
streamlined flow.