linera net up --kubernetes initial version (#1231)
## Motivation

We want to replace `build_and_redeploy`, and also set things up to run end to end tests agaist a locally running Kubernetes cluster.

## Proposal

This is an initial version. We don't have number of validators or numbers of shards being respected in this version, this will be done in a follow up PR.

## Test Plan

Ran it:


![Screenshot 2023-11-10 at 11.17.04.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/HlciHFAoHZW62zn13apJ/6ee4028d-a05d-4a33-913b-912fbf41750a.png)


Then ran `sync-balance` against it, and it succeeded.