[prometheus] Don't log high cardinality labels (#1262)
## Motivation

From the documentation (https://prometheus.io/docs/practices/naming/#labels):
> Do not use labels to store dimensions with high cardinality (many different label values), such as user IDs, email addresses, or other unbounded sets of values.

## Proposal

Oversight on my part, this can cause a huge amount of data to Prometheus, so let's remove it for now. We can figure out another solution if we ever need this level of cardinality. Thanks @ma2bd for pointing out

## Test Plan

Kubernetes CI should fail if these metrics are failing to be logged