Wait for Scylladb initialization instead of failing (#1293)
## Motivation

Right now when we're trying to initialize the shards, we check for ScyllaDB and just fail if it's not ready. That makes Kubernetes go into a `CrashLoopBackoff`, which is an exponential backoff, which will make us wait much more than we need.

## Proposal

Change the script to wait for the initialization instead of failing

## Test Plan

Started clusters locally, a few seconds after ScyllaDB is `Running`, the shards also start being initialized. Initialization felt a lot faster