Persist Grafana dashboards (#1355)
## Motivation

We'll have some dashboards that will contain different data that we're interested in monitoring. To do that we need a way of defining and persisting Grafana dashboards.

## Proposal

Create dashboards in Grafana, export the JSON, add it to our codebase and make sure that JSON gets imported when Grafana is initialized

Infra side PR: https://github.com/linera-io/linera-infra/pull/20

## Test Plan

Deployed locally, saw that the imported dashboard was there