Remove support for reentrant application calls (#1437)
* Keep track of the applications in the call stack

Use a `HashSet` to keep track of which application IDs are present in
the call stack.

* Prevent reentrant application calls

Return an error if it is attempted.

* Remove `reentrant-counter` example

Since reentrant calls are no longer supported, the example does not work
anymore. Also remove the end-to-end test that uses the example.

* Refactor tests to not use a reentrant test app.

Split the `TestApplication` in two, where one is the caller and the
other the callee.

* Don't handle state for cross-application calls

Remove automatic persistence of state before cross-application calls,
which was used as a workaround to allow for reentrant calls.

* Make cross-application call SDK APIs synchronous

They don't need to be `async` anymore because they don't require
interacting with storage.

* Remove unnecessary `async` modifiers in examples

Simplify the examples so that functions aren't asynchronous unless they
need to be.