Fix missing application registration messages (#1486)
* Reword `expect` messages for missing mock apps.

Follow the convention used in other parts of the code.

* Add an `add_message` builder method

Make it easier to schedule a message to be sent to other chains.

* Add a `with_new_session` builder method

Make it easier to request for the creation of new sessions.

* Derive `Clone` for `RawOutgoingMessage`

Make it easier to use in tests.

* Test sending a simple message

Ensure that the receiving chain also receives a system message to
register the receiver application.

* Test messages scheduled during application calls

They should also require registration of the called application if it
schedules a message to be sent.

* Test sending a message from a session

The application that created the session should be registered on the
receiving chain.

* Test sending multiple messages from applications

The receiving chains should receive registrations for the applications
that will handle its received messages.

* Fix registration of called apps that send messages

Ensure all the outgoing messages have an application registered to
handle them on the target chain. This is done by prepending a system
message with the necessary registrations.

Messages can be sent from any of the applications in the call stack, not
just the entrypoint application. Therefore all the outcomes must be
checked for outgoing messages.