Fix deadlock caused by circular references
The `ExecutionStateView` starts a thread to execute the application with
the synchronous runtime, and in its original thread it loops listening
for requests from the application. When the execution finishes, the
thread ends and when the runtime is dropped it also drops the sender of
requests to the `ExecutionStateView`. Therefore the `ExecutionStateView`
then detects that the channel has closed and finishes the loop
continuing execution.

However, with the addition of the `loaded_applications` map, each loaded
application has a strong reference to the runtime, so there's a circular
dependency which prevents the `SyncRuntimeInternal` type from dropping.
This prevents the channel from closing, and causes a deadlock.

To avoid this situation, the `loaded_applications` map must be cleared
at the end of the execution, independently if it succeeded or if it
failed.