Add workspace file with examples (#1544)
## Motivation

For VSCode users, right now `rust-analyzer` doesn't recognize anything in the examples folder. That makes the red squiggles for errors, auto complete, jump to definition, etc not work from within the examples folder.

## Proposal

The reason for that is because examples isn't included in the workspace. I'm adding with this PR a workspace file. If you open your `linera-protocol` checkout using File -> Open Workspace from File and choose this file, it'll open a workspace with both the root directory and the examples directory in it, making all `rust-analyzer` features work on the examples folder as well.

## Test Plan

Tested locally, I see errors in examples now (in my WIP code for another PR)