Create wrappers for registering Prometheus metrics (#1521)
## Motivation

Ideally we would want a way to set a namespace to all our registered metrics

## Proposal

Created wrappers for the metric registration macros we're using, and made sure we set the linera namespace for all of them. Setting the namespace will also add a prefix to the exported metrics. We need to only use these wrappers from now on when registering new metrics, otherwise the namespace won't be properly set

## Test Plan

![Screenshot 2024-01-18 at 14.25.37.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/HlciHFAoHZW62zn13apJ/443eb64f-b459-4dce-a12e-93cb335c4253.png)