Add some metrics related to execution (#1545)
* Add a `MeasureLatency` extension method

Make it simple to measure latency and store it in a histogram. Calling
the method returns a type that automatically finishes the measurement
when it's dropped.

* Refactor how block execution latency is measured

Use the new `MeasureLatency` extension trait.

* Add metric for state hash computation latency

Track it because it is a significant part of the block latency.

* Measure contract and service intantiation latency

Keep track of it to see how it impacts performance.

* Measure message execution latency

Create a histogram of the latency to execute a single message.

* Measure operation execution latency

Create a histogram of the latency to execute a single operation.

* Measure the latency to load contracts and services

Build a histogram of the time to load the bytecodes.