Add a `SharableView` trait (#1641)
* Add a `ClonableView` trait

Allow views to be shared unsafely. This trait replaces `Clone`
implementations for views, and provides a name that brings attention to
the fact that it's not safe unless used carefully.

* Allow deriving `ClonableView`

Implement a procedural derive macro to generate `ClonableView`
implementations.

* Test deriving `SharableView`

Ensure the generated code properly implements `SharableView` for some
test types.

* Create a `TestView` trait and some implementations

Some helper types to simplify writing test cases for different views.

* Test if staged changes are cloned

Ensure that a view's staged changes are properly cloned.

* Test if clone has separate staging area

Ensure that the clone is a separate view instance.