Creating pd-balanced StorageClass for GCP (#1646)
## Motivation

We're now using c3 VM type on GCP, which is incompatible with `pd-standard`

## Proposal

Use `pd-balanced` instead. Local runs can keep using `pd-standard`, which is why I'm not altering anything there. Also setting reclaim policy to `Retain` now, as the default was `Delete` before, which is not what we want.

## Test Plan

Restarted Prometheus pod using this StorageClass on the devnet