Improve support for generics in `linera-views-derive` (#1467)
* Return an `WhereClause` instead of tokens

Make it easier to inspect its predicates so that it can be merged with
any existing where clauses.

* Respect existing where clauses in derive macros

Append new predicates to existing where clauses if possible, instead of
overwriting the existing where clause.

* Add `where_clause` field to `SpecificContextInfo`

Prepare to test with custom `where` clauses.

* Use specific `AngleBracketedGenericArguments` type

Instead of the broader `TokenStream` type. This make it easier to add
more generic arguments to the list.

* Test inclusion of `where` clause in derive input

Ensure it is preserved in the generated implementations.