Remove `wit-bindgen` standalone feature requirement (#1710)
* Rename `result` into `outcome`

Make it consistent with other usages.

* Create some `into_raw` helper conversion methods

Prepare to add an intermediary step from the generic type to the WIT
type, so that the "raw" type (non-generic but also not WIT) can be sent
from the application to the SDK entrypoints.

* Refactor to use `into_raw` conversions

Add the intermediary step and remove the now obsolete `From`
implementations that serialize internal fields.

* Implement contract entrypoint proxy functions

Export some hidden "foreign" functions from the `contract!` macro, and
register them in the `linera-sdk` so that they can be called. The linker
is then able to match them, so that no extra functions are imported or
exported.

* Stub the contract entrypoints in services

Since `linera-sdk` is used indiscriminantly by contracts and services,
the entrypoint proxy functions from one are also imported by the other.
If they are missing, the linker creates additional import entries in the
Wasm file, and runtimes then complain when those imports are requested
but not provided.

* Replace `export_contract!` with implementation

Remove the usage of `wit-bindgen`'s "standalone" feature, which is
deprecated.

* Refactor to move conversion of `QueryContext`

Update the `ServiceStateStorage` to not use the WIT type.

* Implement service entrypoint proxy functions

Export some hidden "foreign" functions from the `service!` macro, and
register them in the `linera-sdk` so that they can be called. The linker
is then able to match them, so that no extra functions are imported or
exported.

* Stub the service entrypoint in contracts

Since `linera-sdk` is used indiscriminantly by contracts and services,
the entrypoint proxy functions from one are also imported by the other.
If they are missing, the linker creates additional import entries in the
Wasm file, and runtimes then complain when those imports are requested
but not provided.

* Replace `export_service!` with implementation

Remove the usage of `wit-bindgen`'s "standalone" feature, which is
deprecated.